+++ sketch_oxide/src/frequency/mod.rs
@@ -12,6 +12,7 @@ pub mod removable_sketch;
 pub mod salsa;
 mod space_saving;
 mod space_saving_pm;
+mod spread_sketch;
 mod tower_sketch;
 
 pub use conservative_count_min::ConservativeCountMin;
@@ -26,4 +27,5 @@ pub use removable_sketch::RemovableUniversalSketch;
 pub use salsa::SALSA;
 pub use space_saving::SpaceSaving;
 pub use space_saving_pm::SpaceSavingPlusMinus;
+pub use spread_sketch::SpreadSketch;
 pub use tower_sketch::TowerSketch;